+++ test/test_helpers/asset_manager_test.rb
@@ -208,4 +208,30 @@ describe GdsApi::TestHelpers::AssetManager do
       end
     end
   end
+
+  describe "#stub_asset_manager_update_asset_forbidden" do
+    describe "when the endpoint is requested with the given ID" do
+      it "raises GdsApi::HTTPForbidden" do
+        asset_id = "some-id"
+        stub_asset_manager_update_asset_forbidden(asset_id)
+
+        assert_raises GdsApi::HTTPForbidden do
+          stub_asset_manager.update_asset(asset_id, {})
+        end
+      end
+    end
+  end
+
+  describe "#stub_asset_manager_update_asset_not_found" do
+    describe "when the endpoint is requested with the given ID" do
+      it "raises GdsApi::HTTPNotFound" do
+        asset_id = "some-id"
+        stub_asset_manager_update_asset_not_found(asset_id)
+
+        assert_raises GdsApi::HTTPNotFound do
+          stub_asset_manager.update_asset(asset_id, {})
+        end
+      end
+    end
+  end
 end