+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   * `stub_asset_manager_create_asset_unprocessable`
   * `stub_asset_manager_request_to_asset_media`
   * `stub_asset_manager_request_to_get_asset`
+  * `stub_asset_manager_update_asset_forbidden`
+  * `stub_asset_manager_update_asset_not_found`
 * Rename `stub_asset_manager_receives_an_asset` test helper to `stub_asset_manager_create_asset`; the old name will be deprecated in a later release ([PR](https://github.com/alphagov/gds-api-adapters/pull/1416))
 
 ## 103.4.0