+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * `stub_asset_manager_update_asset_forbidden`
   * `stub_asset_manager_update_asset_not_found`
 * Rename `stub_asset_manager_receives_an_asset` test helper to `stub_asset_manager_create_asset`; the old name will be deprecated in a later release ([PR](https://github.com/alphagov/gds-api-adapters/pull/1416))
+* Allow `stub_asset_manager_create_asset` test helper to return a body ([PR](https://github.com/alphagov/gds-api-adapters/pull/1416))
 
 ## 103.4.0